agent/requirements.txt
requests
docker
psutil
python-dotenv
